Extract Monix integration into a subproject
